Create Shiftable module
