+++ lib/cipher.rb
@@ -2,10 +2,12 @@ require_relative 'alphabetable'
 require_relative 'key'
 require_relative 'offset'
 require_relative 'positionable'
+require_relative 'shiftable'
 
 class Cipher
   include Alphabetable
   include Positionable
+  include Shiftable
 
   attr_reader :keys,
               :date,
@@ -17,23 +19,7 @@ class Cipher
     @offset = Offset.generate(@date)
   end
 
-  def shift_lookup
-    @shift_lookup ||= positions.zip(@keys.zip(@offset).map(&:sum)).to_h
-  end
-
-  def shift(letter, index)
-    alphabet.zip(alphabet.rotate(shift_lookup[index % 4])).to_h[letter]
-  end
-
-  def shift_new_letter(letter, index)
-    alphabet.include?(letter) ? shift(letter, index) : letter
-  end
-
-  def unshift(letter, index)
-    alphabet.zip(alphabet.rotate(shift_lookup[index % 4] * -1)).to_h[letter]
-  end
-
-  def unshift_new_letter(letter, index)
-    alphabet.include?(letter) ? unshift(letter, index) : letter
+  def shift_values
+    @keys.zip(@offset).map(&:sum)
   end
 end
