+++ lib/encrypt.rb
@@ -5,6 +5,7 @@ require_relative 'enigma'
 require_relative 'key'
 require_relative 'offset'
 require_relative 'positionable'
+require_relative 'shiftable'
 
 handle  = File.open(ARGV[0], 'r')
 message = handle.read.downcase
